+++ test/unit/braintree/graphql/customer_recommendations_payload_spec.js
@@ -4,23 +4,25 @@ const {
   CustomerRecommendations,
   CustomerRecommendationsPayload,
   RecommendedPaymentOption,
-  PaymentOptions,
+  PaymentRecommendation,
 } = require("../../../../lib/braintree/graphql");
 
 describe("CustomerRecommendationsPayload", () => {
-  it("should correctly return isInPayPalNetwork and recommendations", () => {
-    const paymentOptions = new PaymentOptions(
+  it("should correctly return sessionId, isInPayPalNetwork and recommendations", () => {
+    const paymentRecommendations = new PaymentRecommendation(
       RecommendedPaymentOption.PAYPAL,
       1
     );
     const customerRecommendations = new CustomerRecommendations([
-      paymentOptions,
+      paymentRecommendations,
     ]);
     const payload = new CustomerRecommendationsPayload(
+      "session-id",
       true,
       customerRecommendations
     );
 
+    assert.equal(payload.sessionId, "session-id");
     assert.equal(payload.isInPayPalNetwork, true);
     assert.deepStrictEqual(payload.recommendations, customerRecommendations);
   });